+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -1,0 +1,39 @@
+package raisetech.student.management.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourse;
+import raisetech.student.management.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  // 生徒リストをサービスから取得　
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  // 特定の生徒のコースだけをサービスから取得する API を追加
+  @GetMapping("/courseListByStudentId")
+  public List<StudentCourse> getCoursesByStudentId(@RequestParam String studentId) {
+    return service.searchCoursesByStudentId(studentId);
+  }
+
+  // コースリストをサービスから取得
+  @GetMapping("/courseList")
+  public List<StudentCourse> getCourseList() {
+    return service.searchCourseList();
+  }
+}
