+++ src/main/java/raisetech/student/management/repository/StudentRepository.java
@@ -7,7 +7,7 @@ import raisetech.student.management.data.StudentCourse;
 
 /**
  * 受講生情報を扱うリポジトリ。
- *
+ * <p>
  * 全件検索や単一条件での検索、コース情報の検索が行えるクラスです。
  */
 @Mapper
